+++ src/interactive-window/editor-integration/codelensprovider.node.ts
@@ -11,7 +11,7 @@ import {
     IWorkspaceService
 } from '../../platform/common/application/types';
 import { ContextKey } from '../../platform/common/contextKey.node';
-import { disposeAllDisposables } from '../../platform/common/helpers.node';
+import { disposeAllDisposables } from '../../platform/common/helpers';
 import { IFileSystem } from '../../platform/common/platform/types.node';
 
 import { IConfigurationService, IDisposable, IDisposableRegistry } from '../../platform/common/types';
