+++ src/kernels/jupyter/launcher/notebookStarter.node.ts
@@ -12,7 +12,7 @@ import * as uuid from 'uuid/v4';
 import { CancellationError, CancellationToken, Disposable } from 'vscode';
 import { IDisposable } from '@fluentui/react';
 import { Cancellation, createPromiseFromCancellation } from '../../../platform/common/cancellation.node';
-import { disposeAllDisposables } from '../../../platform/common/helpers.node';
+import { disposeAllDisposables } from '../../../platform/common/helpers';
 import { traceInfo, traceError } from '../../../platform/logging';
 import { TemporaryDirectory } from '../../../platform/common/platform/types';
 import { IOutputChannel, Resource } from '../../../platform/common/types';
