+++ src/notebooks/outputs/rendererCommunication.node.ts
@@ -4,7 +4,7 @@
 import { inject, injectable } from 'inversify';
 import { Event, extensions, NotebookEditor, window } from 'vscode';
 import { IExtensionSingleActivationService } from '../../platform/activation/types';
-import { disposeAllDisposables } from '../../platform/common/helpers.node';
+import { disposeAllDisposables } from '../../platform/common/helpers';
 import { IDisposable } from '../../platform/common/types';
 import { noop } from '../../platform/common/utils/misc';
 import { PlotSaveHandler } from './plotSaveHandler.node';
