+++ src/platform/api/kernelApi.node.ts
@@ -10,7 +10,7 @@ import {
     KernelConnectionMetadata as IKernelKernelConnectionMetadata
 } from '../../kernels/types';
 import { INotebookControllerManager } from '../../notebooks/types';
-import { disposeAllDisposables } from '../common/helpers.node';
+import { disposeAllDisposables } from '../common/helpers';
 import { traceInfo } from '../logging';
 import { IDisposable, IDisposableRegistry, IExtensions } from '../common/types';
 import { PromiseChain } from '../common/utils/async';
