+++ src/platform/common/helpers.ts
@@ -1,0 +1,18 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
+'use strict';
+import { IDisposable } from './types';
+
+export function disposeAllDisposables(disposables: IDisposable[] = []) {
+    while (disposables.length) {
+        const disposable = disposables.shift();
+        if (disposable) {
+            try {
+                disposable.dispose();
+            } catch {
+                // Don't care.
+            }
+        }
+    }
+}
